+++ src/main/java/Parser.java
@@ -1,0 +1,68 @@
+public class Parser {
+    public static ToDo generateToDoFromInput(String input) {
+        String[] commands = input.split(" ");
+        String description = "";
+        for (int i = 1; i < commands.length; ++i) {
+            description += commands[i] + " ";
+        }
+        description = description.substring(0, description.length() - 1);
+        return new ToDo(description);
+    }
+
+    public static Deadline generateDeadlineFromInput(String input) {
+        String[] commands = input.split(" ");
+        String description = "";
+        String timeQualifier = "";
+        String timeDescription = "";
+
+        int timeQualifierIndex = 0;
+
+        for (int i = 1; i < commands.length; ++i) {
+            if (commands[i].charAt(0) == '/') {
+                timeQualifierIndex = i;
+                break;
+            }
+            description += commands[i] + " ";
+        }
+        description = description.substring(0, description.length() - 1);
+
+        timeQualifier = commands[timeQualifierIndex].substring(1);
+
+        for (int i = timeQualifierIndex + 1; i < commands.length; ++i) {
+            timeDescription += commands[i] + " ";
+        }
+
+        timeDescription = timeDescription.substring(0, timeDescription.length() - 1);
+
+        return  new Deadline(description, timeQualifier, timeDescription);
+    }
+
+    public static Event generateEventFromInput(String input) {
+        String[] commands = input.split(" ");
+        String description = "";
+        String timeQualifier = "";
+        String timeDescription = "";
+
+        int timeQualifierIndex = 0;
+
+        for (int i = 1; i < commands.length; ++i) {
+            if (commands[i].charAt(0) == '/') {
+                timeQualifierIndex = i;
+                break;
+            }
+            description += commands[i] + " ";
+        }
+        description = description.substring(0, description.length() - 1);
+
+        timeQualifier = commands[timeQualifierIndex].substring(1);
+
+        for (int i = timeQualifierIndex + 1; i < commands.length; ++i) {
+            timeDescription += commands[i] + " ";
+        }
+
+        timeDescription = timeDescription.substring(0, timeDescription.length() - 1);
+
+        return new Event(description, timeQualifier, timeDescription);
+    }
+
+}
