+++ src/main/java/TaskList.java
@@ -1,65 +1,21 @@
-import java.io.File;
-import java.io.FileNotFoundException;
-import java.io.FileWriter;
-import java.io.IOException;
 import java.util.ArrayList;
-import java.util.Scanner;
 
 public class TaskList {
 
     protected static final UI UI = new UI();
     protected ArrayList<Task> taskList;
-    protected String taskDataPath;
 
-    TaskList(String taskDataPath) {
+    TaskList() {
         this.taskList = new ArrayList<Task>();
-        this.taskDataPath = taskDataPath;
-        readTaskData();
+    }
+
+    public ArrayList<Task> getTaskList() {
+        return taskList;
     }
 
     public void add(Task task) {
         taskList.add(task);
         UI.printResponse("Got it. I've added this task:\n" + task.toString() + "\n" + "Now you have " + taskList.size() + " tasks in the list.\n");
-    }
-
-    public void readTaskData() {
-        try {
-            File f = new File(taskDataPath);
-            Scanner s = new Scanner(f);
-            while (s.hasNext()) {
-                String line = s.nextLine();
-                String[] commands = line.split(",");
-
-                Task newTask;
-
-                if (commands[0].equals("T")) {
-                    newTask = new ToDo(commands[2]);
-                } else if (commands[0].equals("D")) {
-                    newTask = new Deadline(commands[2], commands[3], commands[4]);
-                } else {
-                    newTask = new Event(commands[2], commands[3], commands[4]);
-                }
-
-                if(!commands[1].equals("0")) {
-                    newTask.markAsDone();
-                }
-                taskList.add(newTask);
-            }
-        } catch (FileNotFoundException e) {
-            UI.printResponse("I cannot find the data file :(\n");
-        }
-    }
-
-    public void updateTaskData() {
-        try {
-            FileWriter fw = new FileWriter(taskDataPath);
-            for (int i = 0; i < taskList.size(); ++i) {
-                fw.write(taskList.get(i).toWrite());
-            }
-            fw.close();
-        } catch (IOException e) {
-            UI.printResponse("Something went wrong: " + e.getMessage() + "\n");
-        }
     }
 
     public void printTaskList() {
